coreV5/mizore/requirements.txt
toml==0.10.0
redis==3.3.11
Flask==1.0.3
Werkzeug==0.16.0
rq==1.1.0
astropy==3.2.3
rq_dashboard==0.6.1
requests==2.20.0
PyYAML==5.2
